Start decoupling AGP
